+++ packages/extension-core/src/handlers/Extension.ts
@@ -5,8 +5,8 @@ import { db } from "../db"
 import { AccountsHandler } from "../domains/accounts"
 import { AccountType } from "../domains/accounts/types"
 import AppHandler from "../domains/app/handler"
+import { hideGetStartedOnceFunded } from "../domains/app/hideGetStartedOnceFunded"
 import { trackPopupSummaryData } from "../domains/app/popupSummaries"
-import { hideGetStartedOnceFunded } from "../domains/app/utils"
 import { AssetDiscoveryHandler } from "../domains/assetDiscovery/handler"
 import { BalancesHandler } from "../domains/balances"
 import { ChainsHandler } from "../domains/chains"
